requirements.txt
# Agent Framework
langgraph
langchain
langchain-groq

# Groq API
groq

# Code Parsing
tree-sitter
tree-sitter-languages

# Web Framework
fastapi
uvicorn[standard]

# Data handling
pydantic
python-dotenv

# Testing
pytest
# Elasticsearch
elasticsearch==8.11.0